feat() : distance 타입 변경

+++ backend/runIt/src/main/java/com/ssafy/runit/domain/record/entity/Record.java
@@ -31,7 +31,7 @@ public class Record {
      @OneToMany(mappedBy = "record")
      private List<Pace> paceList;
 
-     private double distance;
+     private Integer distance;
 
      private Integer bpm;
